requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
bzip2=1.0.8=h93a5062_5
ca-certificates=2024.2.2=hf0a4a13_0
et-xmlfile=1.1.0=pypi_0
libexpat=2.5.0=hb7217d7_1
libffi=3.4.2=h3422bc3_5
libsqlite=3.45.1=h091b4b1_0
libzlib=1.2.13=h53f4e23_5
ncurses=6.4=h463b476_2
numpy=1.26.4=pypi_0
openpyxl=3.1.2=pypi_0
openssl=3.2.1=h0d3ecfb_0
pandas=2.2.0=pypi_0
pip=24.0=pyhd8ed1ab_0
python=3.12.2=hdf0ec26_0_cpython
python-dateutil=2.8.2=pypi_0
pytz=2024.1=pypi_0
readline=8.2=h92ec313_1
setuptools=69.1.0=pyhd8ed1ab_1
six=1.16.0=pypi_0
tk=8.6.13=h5083fa2_1
tzdata=2024.1=pypi_0
wheel=0.42.0=pyhd8ed1ab_0
xz=5.2.6=h57fd34a_0
